python_server/requirements.txt
mcp[cli]
httpx
python-dotenv
openai
